Worked on Shield, mystery and penetrate powerups. dont have time to test them if somebody else will.

+++ EnemyShip1.java
@@ -67,7 +67,7 @@ public class EnemyShip1 extends EnemyShip
     {
         if(hitee.owner instanceof GoodShip)
         {
-            hitee.delete();
+            if(!Projectile.penetrate)hitee.delete();
             GoodShip killer = (GoodShip) (hitee.owner);
             addPowerup();
             getWorld().removeObject(this);
