+++ EnemyShip4.java
@@ -70,9 +70,16 @@ public class EnemyShip4 extends EnemyShip
     {
         if(hitee.owner instanceof GoodShip)
         {
-            hitee.delete();
+            if(!Projectile.penetrate)hitee.delete();
             if(!shielded)
             {
+                GoodShip killer = (GoodShip) (hitee.owner);
+                addPowerup();
+                getWorld().removeObject(this);
+                killer.score = killer.score + 500;
+                killer.enemiesKilled++;
+            }
+            else if(Projectile.penetrate){
                 GoodShip killer = (GoodShip) (hitee.owner);
                 addPowerup();
                 getWorld().removeObject(this);
